+++ docs/api/changelog.md
@@ -4,6 +4,11 @@
     .container hr{margin-top: 5px; }
 </style>
 
+## Mar 5, 2019
+
+- Remove endpoint: publish a post via `POST /api/posts/:post_id/publish` ([@GeekyShacklebolt])
+- Add endpoint: unapprove a post via `POST /api/posts/:post_id/unapprove` 
+
 ## Dec 15, 2018
 - Add an app for `posts`. ([@realslimshanky])
   This adds following features: